requirements.txt
anthropic==0.25.0 
chromadb==0.4.24 
sentence-transformers==2.7.0 
pdfplumber==0.11.0 
openpyxl==3.1.2 
pandas==2.2.1 
streamlit==1.33.0 
langchain-text-splitters==0.0.2 
python-dotenv==1.0.1 
torch==2.2.2 
loguru==0.7.2  
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
uuid==1.30